Pin duplicate generic requires diagnostic schema
