+++ docs/V1_SPEC.md
@@ -366,7 +366,7 @@ Generic behavior inheritance with child type-parameter parent args is covered by
 | README and contributor truth assertions | implemented | `tests/docs_truth.rs` |
 | Strict resolver, symbol IDs, privacy | implemented | Resolver/module/privacy tests |
 | AST/symbols JSON emission | constrained | `emit_json_ast_command_outputs_resolved_module_graph` checks unchecked AST module-graph output, `emit_json_ast_module_graph_schema_matches_golden` pins the unchecked AST module graph schema, `emit_json_symbols_command_outputs_module_symbol_tables` checks resolved symbol table output, `emit_json_symbols_module_graph_schema_matches_golden` pins the resolved symbols module graph schema, `emit_json_symbols_generic_behavior_association_schema_matches_golden` pins generic behavior method signatures and `Point` association metadata for `Json<StaticString>`, `emit_json_ast_rejects_hand_authored_json_before_unchecked_ir_override`, `emit_json_symbols_rejects_hand_authored_json_before_resolver_override` |
-| Diagnostics JSON emission | constrained | `emit_json_diagnostics_command_outputs_machine_readable_errors` checks machine-readable errors, `emit_json_diagnostics_removed_return_schema_matches_golden` pins removed-return suggested fix schema, `emit_json_diagnostics_behavior_derive_gate_schema_matches_golden` pins feature-gate context for reserved generated behavior association, `emit_json_diagnostics_generic_association_gate_schema_matches_golden` pins feature-gate context for reserved generic behavior association targets, `emit_json_diagnostics_generic_result_method_arity_schema_matches_golden` pins a hard generic `Result<T, E>` method arity diagnostic without followups, `emit_json_diagnostics_generic_result_method_bound_schema_matches_golden` pins a hard generic behavior-bound diagnostic without method-body followups, `emit_json_diagnostics_generic_result_method_inference_schema_matches_golden` pins a hard generic inference conflict without argument/return followups, `emit_json_diagnostics_generic_behavior_overlap_schema_matches_golden` pins a behavior implementation coherence diagnostic for overlapping generic parent/child impls, `emit_json_diagnostics_generic_requires_missing_impl_schema_matches_golden` pins an explicit `.requires` missing implementation diagnostic for `Json<StaticString>`, `emit_json_diagnostics_rejects_hand_authored_json_before_diagnostic_override`; `docs/DIAGNOSTICS.md` catalogs the currently JSON-stable public codes, while broader diagnostic-code coverage is still required |
+| Diagnostics JSON emission | constrained | `emit_json_diagnostics_command_outputs_machine_readable_errors` checks machine-readable errors, `emit_json_diagnostics_removed_return_schema_matches_golden` pins removed-return suggested fix schema, `emit_json_diagnostics_behavior_derive_gate_schema_matches_golden` pins feature-gate context for reserved generated behavior association, `emit_json_diagnostics_generic_association_gate_schema_matches_golden` pins feature-gate context for reserved generic behavior association targets, `emit_json_diagnostics_generic_result_method_arity_schema_matches_golden` pins a hard generic `Result<T, E>` method arity diagnostic without followups, `emit_json_diagnostics_generic_result_method_bound_schema_matches_golden` pins a hard generic behavior-bound diagnostic without method-body followups, `emit_json_diagnostics_generic_result_method_inference_schema_matches_golden` pins a hard generic inference conflict without argument/return followups, `emit_json_diagnostics_generic_behavior_overlap_schema_matches_golden` pins a behavior implementation coherence diagnostic for overlapping generic parent/child impls, `emit_json_diagnostics_generic_requires_missing_impl_schema_matches_golden` pins an explicit `.requires` missing implementation diagnostic for `Json<StaticString>`, `emit_json_diagnostics_duplicate_generic_requires_schema_matches_golden` pins duplicate `.requires` resolver diagnostics for `Json<StaticString>`, `emit_json_diagnostics_rejects_hand_authored_json_before_diagnostic_override`; `docs/DIAGNOSTICS.md` catalogs the currently JSON-stable public codes, while broader diagnostic-code coverage is still required |
 | Typed JSON emission | constrained | `emit_json_typed_command_outputs_checked_program` checks `semantic_status: "checked"`, `emit_json_typed_generic_method_schema_matches_golden` pins generic method specialization for `Box_i32` / `Box.get_i32`, `emit_json_typed_generic_option_schema_matches_golden` pins concrete `Option<i32>` enum specialization and typed `unwrap_or_i32` calls, `emit_json_typed_generic_result_schema_matches_golden` pins concrete `Result<i32, StaticString>` enum specialization and typed `unwrap_or_i32_StaticString` calls, `emit_json_typed_nested_generic_result_schema_matches_golden` pins nested `Result<Option<i32>, StaticString>` and dependent `Option<i32>` typed payloads/calls, `emit_json_typed_generic_behavior_association_schema_matches_golden` pins concrete generic behavior association dispatch to `Point.encode__Json_StaticString`, `emit_json_typed_generic_behavior_bound_ufcs_schema_matches_golden` pins substituted generic behavior-bound UFCS dispatch from `encode_Point` to `Point.encode__Json_Point`, `emit_json_typed_rejects_hand_authored_json_before_checked_ir_override`; broader typed JSON schemas still required |
 | HIR JSON emission | constrained | `emit_json_hir_outputs_checked_declaration_graph` checks `schema_version: 0`, `emit_json_hir_outputs_enum_function_and_global_declarations` covers enum variants/payloads, function params/returns, and globals, `emit_json_hir_declaration_schema_matches_golden` pins the checked declaration schema, `emit_json_hir_generic_option_schema_matches_golden` pins concrete `Option<i32>` enum/function specialization output, `emit_json_hir_generic_result_schema_matches_golden` pins concrete `Result<i32, StaticString>` enum/function specialization output, `emit_json_hir_nested_generic_result_schema_matches_golden` pins nested `Result<Option<i32>, StaticString>` and dependent `Option<i32>` specialization output, `emit_json_hir_generic_behavior_association_schema_matches_golden` pins concrete generic behavior association function signatures, `emit_json_hir_generic_behavior_bound_ufcs_schema_matches_golden` pins generic behavior-bound UFCS declaration output for `encode_Point` and `Point.encode__Json_Point`, `emit_json_hir_rejects_hand_authored_json_before_ir_override`; broader golden tests still required |
 | MIR JSON emission | constrained | `emit_json_mir_outputs_checked_minimal_function_graph` checks `schema_version: 0`, `emit_json_mir_minimal_function_schema_matches_golden` pins the checked minimal function/block/terminator schema, `emit_json_mir_outputs_match_arm_schema` covers match kind, arm patterns/bindings, and block results, `emit_json_mir_match_schema_matches_golden` pins the checked match schema, `emit_json_mir_generic_option_schema_matches_golden` pins concrete `Option<i32>` enum construction and match lowering, `emit_json_mir_generic_result_schema_matches_golden` pins concrete `Result<i32, StaticString>` enum construction and match lowering, `emit_json_mir_nested_generic_result_schema_matches_golden` pins nested `Result<Option<i32>, StaticString>` construction, calls, and match lowering, `emit_json_mir_generic_behavior_association_schema_matches_golden` pins concrete generic behavior association call lowering, `emit_json_mir_generic_behavior_bound_ufcs_schema_matches_golden` pins lowered generic behavior-bound UFCS dispatch from `encode_Point` to `Point.encode__Json_Point`, `emit_json_mir_rejects_hand_authored_json_before_ir_override`; broader golden tests still required |
