+++ tests/docs_truth/public_docs.rs
@@ -270,6 +270,8 @@ fn diagnostics_catalog_documents_json_stable_codes() {
         "removed syntax or reserved syntax",
         "replace_removed_return_with_final_expression",
         "feature_gate",
+        "E3500",
+        "resolver validation failure for duplicate explicit type association requirements",
         "E5000",
         "generic inference conflict",
         "E5001",
@@ -281,6 +283,7 @@ fn diagnostics_catalog_documents_json_stable_codes() {
         "E6010",
         "behavior implementation coherence failure",
         "tests/fixtures/ir_json/diagnostics_return.golden.json",
+        "tests/fixtures/ir_json/diagnostics_duplicate_generic_requires.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_result_method_arity.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_result_method_bound.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_requires_missing_impl.golden.json",
