+++ tests/integration/cli_build/frontend_json.rs
@@ -1,6 +1,8 @@
 use crate::support::*;
 use std::process::Command;
 
+#[path = "frontend_json/diagnostics_behavior_association_golden.rs"]
+mod diagnostics_behavior_association_golden;
 #[path = "frontend_json/diagnostics_golden.rs"]
 mod diagnostics_golden;
 #[path = "frontend_json/diagnostics_json.rs"]
