+++ tests/integration/cli_build/frontend_json/diagnostics_behavior_association_golden.rs
@@ -1,0 +1,121 @@
+use std::path::Path;
+use std::process::Command;
+
+fn fixture(path: &str) -> std::path::PathBuf {
+    Path::new(env!("CARGO_MANIFEST_DIR")).join(path)
+}
+
+#[test]
+fn emit_json_diagnostics_generic_requires_missing_impl_schema_matches_golden() {
+    let tmp = tempfile::tempdir().expect("create temp dir");
+    let zen_path = tmp.path().join("generic_requires_missing_impl.zen");
+    std::fs::write(
+        &zen_path,
+        r#"
+Point: { x: i32 }
+
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point.requires(Json<StaticString>)
+
+main = () i32 {
+    0
+}
+"#,
+    )
+    .expect("write generic requires missing impl source");
+
+    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
+        .args(["emit-json", "diagnostics", zen_path.to_str().unwrap()])
+        .output()
+        .expect("run zen emit-json diagnostics");
+
+    assert!(
+        !output.status.success(),
+        "zen emit-json diagnostics should fail on missing generic behavior requires impl: stdout={}, stderr={}",
+        String::from_utf8_lossy(&output.stdout),
+        String::from_utf8_lossy(&output.stderr)
+    );
+
+    let actual = String::from_utf8(output.stdout).expect("diagnostics stdout is UTF-8");
+    let json: serde_json::Value =
+        serde_json::from_str(&actual).expect("diagnostics stdout is JSON");
+    assert_eq!(
+        json["diagnostics"]
+            .as_array()
+            .expect("diagnostics array")
+            .len(),
+        1,
+        "missing generic behavior requires impl should emit one diagnostic: {json}"
+    );
+
+    let normalized = actual.replace(tmp.path().to_str().expect("tmp path is UTF-8"), "$TMP");
+    let expected_path =
+        fixture("tests/fixtures/ir_json/diagnostics_generic_requires_missing_impl.golden.json");
+    let expected = std::fs::read_to_string(&expected_path)
+        .unwrap_or_else(|err| panic!("read {}: {err}", expected_path.display()));
+
+    assert_eq!(normalized.trim(), expected.trim());
+}
+
+#[test]
+fn emit_json_diagnostics_duplicate_generic_requires_schema_matches_golden() {
+    let tmp = tempfile::tempdir().expect("create temp dir");
+    let zen_path = tmp.path().join("duplicate_generic_requires.zen");
+    std::fs::write(
+        &zen_path,
+        r#"
+Point: { x: i32 }
+
+Json<T>: behavior {
+    encode: (Self) T
+}
+
+Point.implements(Json<StaticString>) {
+    encode = (value: Point) StaticString { "point" }
+}
+
+Point.requires(Json<StaticString>)
+Point.requires(Json<StaticString>)
+
+main = () i32 {
+    0
+}
+"#,
+    )
+    .expect("write duplicate generic requires source");
+
+    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
+        .args(["emit-json", "diagnostics", zen_path.to_str().unwrap()])
+        .output()
+        .expect("run zen emit-json diagnostics");
+
+    assert!(
+        !output.status.success(),
+        "zen emit-json diagnostics should fail on duplicate generic behavior requires: stdout={}, stderr={}",
+        String::from_utf8_lossy(&output.stdout),
+        String::from_utf8_lossy(&output.stderr)
+    );
+
+    let actual = String::from_utf8(output.stdout).expect("diagnostics stdout is UTF-8");
+    let json: serde_json::Value =
+        serde_json::from_str(&actual).expect("diagnostics stdout is JSON");
+    assert_eq!(
+        json["diagnostics"]
+            .as_array()
+            .expect("diagnostics array")
+            .len(),
+        1,
+        "duplicate generic behavior requires should emit one diagnostic: {json}"
+    );
+
+    let normalized = actual.replace(tmp.path().to_str().expect("tmp path is UTF-8"), "$TMP");
+    let expected_path =
+        fixture("tests/fixtures/ir_json/diagnostics_duplicate_generic_requires.golden.json");
+    let expected = std::fs::read_to_string(&expected_path)
+        .unwrap_or_else(|err| panic!("read {}: {err}", expected_path.display()));
+
+    assert_eq!(normalized.trim(), expected.trim());
+}
